Don't depend on the initial execution order of the threads.
The first iteration of the t2 instance of threadx() is unlocked.  There
is no guarantee that it will have completely initialized func_set before
thread1() tries to use it.  Fix this by moving all of the intialization
to main(), before any of the threads are spawned.

Skip all of the cpu pinning code on FreeBSD.